PCHR-2863: Throw exception in NotificationSender when from email is not configured.
