+++ uk.co.compucorp.civicrm.hrleaveandabsences/CRM/HRLeaveAndAbsences/Service/LeaveRequestMailNotificationSender.php
@@ -2,6 +2,7 @@
 
 use CRM_HRLeaveAndAbsences_Mail_Message as Message;
 use CRM_Core_BAO_MessageTemplate as MessageTemplate;
+use CRM_HRLeaveAndAbsences_Exception_InvalidLeaveRequestMailNotificationSenderException as InvalidLeaveRequestMailNotificationSenderException;
 
 class CRM_HRLeaveAndAbsences_Service_LeaveRequestMailNotificationSender {
 
@@ -13,14 +14,21 @@ class CRM_HRLeaveAndAbsences_Service_LeaveRequestMailNotificationSender {
    * @return array
    *  an array containing the status of the mail sent for each recipient
    *  ['test@example.com' => true, 'tester@example.com' => true]
+   *
+   * @throws InvalidLeaveRequestMailNotificationSenderException
    */
   public function send(Message $message) {
-
     $recipientEmails = $message->getRecipientEmails();
     $templateID = $message->getTemplateID();
     $contactID = $message->getLeaveContactID();
     $fromEmail = $message->getFromEmail();
     $status = [];
+
+    if(is_null($fromEmail)) {
+      throw new InvalidLeaveRequestMailNotificationSenderException(
+        'From Email Address need to be configured in order to allow Email notifications'
+      );
+    }
 
     foreach($recipientEmails as $recipient) {
       $recipientID = $recipient['api.Contact.get']['values'][0]['id'];
